% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apple}
\alias{apple}
\title{French fruit producers, 1984-1986}
\format{
A data frame with 405 observations on the following 10 variables:
\describe{
\item{Time}{Period of observation: 1984-1986.}
\item{ID}{Id of the producer.}
\item{K}{Capital (including land).}
\item{L}{Labour.}
\item{M}{Materials.}
\item{Y1Z}{Apples production.}
\item{Y2Z}{Alternative products.}
\item{PK}{Capital price.}
\item{PL}{Labour price.}
\item{PM}{Materials price.}
}
}
\source{
\url{http://qed.econ.queensu.ca/jae/1996-v11.6/ivaldi-ladoux-ossard-simioni/}
}
\description{
This data set from the Réseau d'Information Comptable Agricole (RICA), and
 Institut National de la Recherche Agronomique (INRA) contains agriculture's
 input and output quantities for 184 French apple producers between 1984 and
 1986. Input prices are also provided.
}
\details{
Further details on the data and the different variables can be found
in the reference.
}
\examples{
str(apple)
summary(apple)
}
\references{
Ivaldi, M., Ladoux, N., Ossard, H., & Simioni, M. (1996). Comparing Fourier
and translog specifications of multiproduct technology: Evidence from an
incomplete panel of French farmers. \emph{Journal of Applied Econometrics},
\bold{11}(6), 649-667.
doi: 10.1002/(Sici)1099-1255(199611)11:6<649::Aid-Jae416>3.0.Co;2-4
}
\keyword{datasets}
