% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costEffDecomp.R
\name{costEffDecomp}
\alias{costEffDecomp}
\title{Cost efficiency decomposition functions for nonconvex technologies}
\usage{
costEffDecomp(
  xobs,
  yobs,
  wobs,
  xref = NULL,
  yref = NULL,
  parallel = FALSE,
  cores = availableCores(omit = 1)
)
}
\arguments{
\item{xobs}{a \eqn{K \times N} matrix with \eqn{K} the number of inputs and
\eqn{N} the number of decision making units (DMUs) under evaluation.}

\item{yobs}{a \eqn{Q \times N} matrix with \eqn{Q} the number of outputs and
\eqn{N} the number of DMUs under evaluation.}

\item{wobs}{a \eqn{K \times N} matrix of input prices.}

\item{xref}{(optional). a  \eqn{K \times N_r} matrix on inputs of DMUs
serving as benchmark (or use to define the technology).
Default to \code{'xobs'}.}

\item{yref}{(optional). a  \eqn{Q \times N_r} matrix on outputs of DMUs
serving as benchmark (or use to define the technology).
Default to \code{'yobs'}.}

\item{parallel}{Logical. Allows parallel computation. If \code{FALSE}
(default), the estimation is conducted in sequential mode. If \code{TRUE},
parallel mode is activated using the number of cores specified in
\code{cores}. When the sample size is small, it is recommended to keep the
\code{parallel} option to its default value (\code{FALSE}).}

\item{cores}{Integer. Used only if \code{parallel = TRUE}. It specifies the
number of cores to be used for parallel computation. By default,
\code{cores = availableCores(omit = 1)}.}
}
\value{
\code{costEffDecomp} returns a tibble that contains five columns.
The first column (\code{'TE'}), contains the technical efficiency scores
obtained under variable returns to scale - vrs, the second column (\code{'SCE'})
contains the scale efficiency, the third column (\code{'OTE'} contains the
overall technical efficiency obtained under constant returns to scale - crs -
the fourth column (\code{'AE'} contains the allocative efficiency, and the
fifth column (\code{'OE'}) contains the overall efficiency (cost efficiency)
obtained under crs. The following equalities are true:
\eqn{OTE = TE \times SCE} and \eqn{OE = AE \times OTE}
}
\description{
Given a vector of input prices, \code{costEffDecomp} decomposes the overall
efficiency (cost efficiency) into several efficiency components.
}
\details{
All estimations are based on the implicit enumeration algorithm
(Briec et al. 2004).
}
\examples{
fdh_cost_decomp <- costEffDecomp(xobs = as.matrix(apple[, c('K', 'L', 'M')]),
yobs = as.matrix(apple[, c('Y1Z', 'Y2Z')]),
wobs = as.matrix(apple[, c('PK', 'PL', 'PM')]))
}
\references{
Briec, W., Kerstens, K., & Eeckaut, P. V. (2004). Non-convex
technologies and cost functions: Definitions, duality and nonparametric
tests of convexity. \emph{Journal of Economics-Zeitschrift Fur Nationalokonomie},
\bold{81}(2), 155-192. doi: \url{10.1007/s00712-003-0620-y}
}
\keyword{models}
\keyword{optimize}
