% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distFun.R
\name{distFun}
\alias{distFun}
\title{Farrell distance functions for nonconvex technologies}
\usage{
distFun(
  xobs,
  yobs,
  xref = NULL,
  yref = NULL,
  rts = "vrs",
  orientation = "out",
  parallel = FALSE,
  cores = availableCores(omit = 1)
)
}
\arguments{
\item{xobs}{a \eqn{K \times N} matrix with \eqn{K} the number of inputs and
\eqn{N} the number of decision making units (DMUs) under evaluation.}

\item{yobs}{a \eqn{Q \times N} matrix with \eqn{Q} the number of outputs and
\eqn{N} the number of DMUs under evaluation.}

\item{xref}{(optional). a  \eqn{K \times N_r} matrix on inputs of DMUs
serving as benchmark (or use to define the technology). Default to \code{'xobs'}.}

\item{yref}{(optional). a  \eqn{Q \times N_r} matrix on outputs of DMUs
serving as benchmark (or use to define the technology). Default to \code{'yobs'}.}

\item{rts}{Character string specifying the returns to scale assumption.
The default value is \code{'vrs'} (variable returns to scale). Other possible
options are \code{'crs'} (constant returns to scale), \code{'nirs'}
(non-increasing returns to scale), or \code{'ndrs'} (non-decreasing returns
to scale).}

\item{orientation}{Character string specifying the orientation.
The default value is \code{'out'} for output-orientation. Other possibilities
are \code{'in'} for input-orientation, and \code{'graph'} for graph
orientation.}

\item{parallel}{Logical. Allows parallel computation. If \code{FALSE}
(default), the estimation is conducted in sequential mode. If \code{TRUE},
parallel mode is activated using the number of cores specified in
\code{cores}. When the sample size is small, it is recommended to keep the
\code{parallel} option to its default value (\code{FALSE}).}

\item{cores}{Integer. Used only if \code{parallel = TRUE}. It specifies the
number of cores to be used for parallel computation. By default,
\code{cores = availableCores(omit = 1)}.}
}
\value{
\code{distFun} returns a tibble that contains two columns. The first
column (\code{'Efficiencies'}), contains the Farrell efficiency scores,
while the second column (\code{'Benchmarks'}) contains the benchmark for
each evaluated observation.
}
\description{
Input-, Output- and graph-oriented radial efficiency measures are computed
assuming nonconvex production technologies.
}
\details{
All efficiency measures are computed using implicit enumeration
(Briec et al. 2004, Briec and Kerstens 2006). Let's assume that there are
\eqn{1,\cdots, N} decision making units (DMUs). Each DMU uses \eqn{K} inputs
to produce \eqn{Q} outputs. The enumeration algorithms to obtain the Farrell
efficiency scores are detailed next.

For the \strong{input orientation} we have the following:

\itemize{
\item{variable returns to scale - \code{'vrs'}}{\deqn{E_i\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right)=\underset{\mathbf{y}^\circ \leq \mathbf{y}_n}{\min}
\underset{k=1,\cdots, K}{\max}\left\{\frac{x_{kn}}{x_k^\circ}\right\}}}

\item{constant returns to scale - \code{'crs'}}{\deqn{E_i\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right) = \underset{n=1, \cdots, N}{\min}
\left\{\underset{k=1,\cdots, K}{\max}\frac{x_{kn}}{x_k^\circ}\cdot\underset{q=1,
\cdots, Q}{\max}\frac{y_q^\circ}{y_{qn}}\right\}}}

\item{nonincreasing returns to scale - \code{'nirs'}}{\deqn{E_i\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right)=\underset{\mathbf{y}^\circ \leq \mathbf{y}_n}{\min}
\left\{\underset{k=1,\cdots, K}{\max}\frac{x_{kn}}{x_k^\circ}\cdot\underset{q=1,
\cdots, Q}{\max}\frac{y_q^\circ}{y_{qn}}\right\}}}

\item{nondecreasing returns to scale - \code{'ndrs'}}{\deqn{E_i\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right)=\underset{n=1, \cdots, N}{\min}
\left\{\underset{k=1,\cdots, K}{\max}\frac{x_{kn}}{x_k^\circ}\cdot \max \left\{\underset{q=1,
\cdots, Q}{\max}\frac{y_q^\circ}{y_{qn}}, 1\right\}\right\}}}
}

In the case of the \strong{output orientation}, we have:

\itemize{
\item{variable returns to scale - \code{'vrs'}}{\deqn{E_o\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right)=\underset{\mathbf{x}^\circ \geq \mathbf{x}_n}{\max}
\underset{q=1,\cdots, Q}{\min}\left\{\frac{y_{qn}}{y_q^\circ}\right\}}}

\item{constant returns to scale - \code{'crs'}}{\deqn{E_o\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right) = \underset{n=1, \cdots, N}{\max}
\left\{\underset{q=1,\cdots, Q}{\min}\frac{y_{qn}}{y_q^\circ}\cdot\underset{k=1,
\cdots, K}{\min}\frac{x_k^\circ}{x_{kn}}\right\}}}

\item{nonincreasing returns to scale - \code{'nirs'}}{\deqn{E_o\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right)=\underset{n=1, \cdots, N}{\max}
\left\{\underset{q=1,\cdots, Q}{\min}\frac{y_{qn}}{y_q^\circ}\cdot \min \left\{\underset{k=1,
\cdots, K}{\min}\frac{x_k^\circ}{x_{kn}}, 1\right\}\right\}}}

\item{nondecreasing returns to scale - \code{'ndrs'}}{\deqn{E_o\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right)=\underset{\mathbf{x}^\circ \geq \mathbf{x}_n}{\max}
\left\{\underset{q=1,\cdots, Q}{\min}\frac{y_{qn}}{y_q^\circ}\cdot\underset{k=1,
\cdots, K}{\min}\frac{x_k^\circ}{x_{kn}}\right\}}}
}


Finally, for the \strong{graph orientation}, let's first define:

\deqn{\delta = \left(\underset{q=1, \cdots, Q}{\max}
\frac{y_q^\circ}{y_{qn}} \cdot \underset{k=1,\cdots, K}{\min}\frac{x_k^\circ}{x_{kn}}\right)^{1/2}}

The enumeration algorithms summarize as follows:

\itemize{
\item{variable returns to scale - \code{'vrs'}}{\deqn{E_{GR}\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right)=\underset{n=1, \cdots, N}{\min} \left\{\max
\left(\underset{k=1,\cdots, K}{\max}\frac{x_{kn}}{x_k^\circ},
\underset{q=1,\cdots, Q}{\max}\frac{y_q^\circ}{y_{qn}}\right)\right\}}}

\item{constant returns to scale - \code{'crs'}}{\deqn{E_{GR}\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right) = \underset{n=1, \cdots, N}{\min}
\left\{\underset{q=1,
\cdots, Q}{\max}\frac{y_q^\circ}{y_{qn}}\Big/\underset{k=1,\cdots, K}{\min}\frac{x_k^\circ}{x_{kn}}\right\}^{1/2}}}

\item{nonincreasing returns to scale - \code{'nirs'}}{ \cr
-If \eqn{\delta \leq 1} then
\deqn{E_{GR}\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right) = \underset{n=1, \cdots, N}{\min}
\left\{\underset{q=1,
\cdots, Q}{\max}\frac{y_q^\circ}{y_{qn}}\Big/\underset{k=1,\cdots, K}{\min}\frac{x_k^\circ}{x_{kn}}\right\}^{1/2}}
-Else
\deqn{E_{GR}\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right) = \underset{n=1, \cdots, N}{\min}\left\{\max\left[\underset{q=1, \cdots, Q}{\max}
\frac{y_q^\circ}{y_{qn}}, \left(\underset{k=1,\cdots, K}{\min}\frac{x_k^\circ}{x_{kn}}\right)^{-1} \right]\right\}}
}

\item{nondecreasing returns to scale - \code{'ndrs'}}{ \cr
-If \eqn{\delta \geq 1} then
\deqn{E_{GR}\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right) = \underset{n=1, \cdots, N}{\min}
\left\{\underset{q=1,
\cdots, Q}{\max}\frac{y_q^\circ}{y_{qn}}\Big/\underset{k=1,\cdots, K}{\min}\frac{x_k^\circ}{x_{kn}}\right\}^{1/2}}
-Else
\deqn{E_{GR}\left(\mathbf{x}^\circ,
\mathbf{y}^\circ\right) = \underset{n=1, \cdots, N}{\min}\left\{\max\left[\underset{q=1, \cdots, Q}{\max}
\frac{y_q^\circ}{y_{qn}}, \left(\underset{k=1,\cdots, K}{\min}\frac{x_k^\circ}{x_{kn}}\right)^{-1} \right]\right\}}
}}
}
\examples{
# Input orientation - variable returns to scale
fdh_in_vrs <- distFun(xobs = as.matrix(apple[, c('K', 'L', 'M')]),
yobs = as.matrix(apple[, c('Y1Z', 'Y2Z')]), rts = 'vrs', orientation = 'in')

summary(fdh_in_vrs[, 1])

# Output orientation - constant returns to scale
fdh_out_crs <- distFun(xobs = as.matrix(apple[, c('K', 'L', 'M')]),
yobs = as.matrix(apple[, c('Y1Z', 'Y2Z')]), rts = 'crs', orientation = 'out')

summary(fdh_out_crs[, 1])

# Graph orientation - non-decreasing returns to scale
fdh_graph_ndrs <- distFun(xobs = as.matrix(apple[, c('K', 'L', 'M')]),
yobs = as.matrix(apple[, c('Y1Z', 'Y2Z')]), rts = 'ndrs', orientation = 'graph')

summary(fdh_graph_ndrs[, 1])
}
\references{
Briec, W., Kerstens, K., & Eeckaut, P. V. (2004). Non-convex
technologies and cost functions: Definitions, duality and nonparametric
tests of convexity. \emph{Journal of Economics-Zeitschrift Fur Nationalokonomie},
\bold{81}(2), 155-192. doi: \url{10.1007/s00712-003-0620-y}

Briec, W., & Kerstens, K. (2006). Input, output and graph technical
efficiency measures on non-convex FDH models with various scaling laws:
An integrated approach based upon implicit enumeration algorithms. \emph{Top},
\bold{14}(1), 135-166. doi: \url{10.1007/BF02579006}
}
\keyword{models}
\keyword{optimize}
