% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revEffDecomp.R
\name{revEffDecomp}
\alias{revEffDecomp}
\title{Revenue efficiency decomposition functions for nonconvex technologies}
\usage{
revEffDecomp(
  xobs,
  yobs,
  pobs,
  xref = NULL,
  yref = NULL,
  parallel = FALSE,
  cores = availableCores(omit = 1)
)
}
\arguments{
\item{xobs}{a \eqn{K \times N} matrix with \eqn{K} the number of inputs and
\eqn{N} the number of decision making units (DMUs) under evaluation.}

\item{yobs}{a \eqn{Q \times N} matrix with \eqn{Q} the number of outputs and
\eqn{N} the number of DMUs under evaluation.}

\item{pobs}{a \eqn{Q \times N} matrix of output prices.}

\item{xref}{(optional). a  \eqn{K \times N_r} matrix on inputs of DMUs
serving as benchmark (or use to define the technology).
Default to \code{'xobs'}.}

\item{yref}{(optional). a  \eqn{Q \times N_r} matrix on outputs of DMUs
serving as benchmark (or use to define the technology).
Default to \code{'yobs'}.}

\item{parallel}{Logical. Allows parallel computation. If \code{FALSE}
(default), the estimation is conducted in sequential mode. If \code{TRUE},
parallel mode is activated using the number of cores specified in
\code{cores}. When the sample size is small, it is recommended to keep the
\code{parallel} option to its default value (\code{FALSE}).}

\item{cores}{Integer. Used only if \code{parallel = TRUE}. It specifies the
number of cores to be used for parallel computation. By default,
\code{cores = availableCores(omit = 1)}.}
}
\value{
\code{revEffDecomp} returns a tibble that contains five columns.
The first column (\code{'TE'}), contains the technical efficiency scores
obtained under variable returns to scale - vrs, the second column (\code{'SCE'})
contains the scale efficiency, the third column (\code{'OTE'} contains the
overall technical efficiency obtained under constant returns to scale - crs -
the fourth column (\code{'AE'} contains the allocative efficiency, and the
fifth column (\code{'OE'}) contains the overall efficiency (cost efficiency)
obtained under crs. The following equalities are true:
\eqn{OTE = TE \times SCE} and \eqn{OE = AE \times OTE}
}
\description{
Given a vector of output prices, \code{revEffDecomp} decomposes the overall
efficiency (revenue efficiency) into several efficiency components.
}
\details{
All estimations are based on the implicit enumeration algorithm
}
\examples{
\dontrun{
fdh_rev_decomp <- revEffDecomp(xobs = as.matrix(usagri[, c('q.capital',
'q.land', 'q.labor', 'q.materials')]),
yobs = as.matrix(usagri[, c('q.livestock', 'q.crops', 'q.otherout')]),
pobs = as.matrix(usagri[, c('p.livestock', 'p.crops', 'p.otherout')]))
}
}
\keyword{models}
\keyword{optimize}
